SubjectType 클래스 추가, 코드 수정
