+++ camp-management-app/src/camp/model/Subject.java
@@ -5,9 +5,9 @@ import java.util.Objects;
 public class Subject {
     private String subjectId;
     private String subjectName;
-    private String subjectType;
+    private SubjectType subjectType;
 
-    public Subject(String seq, String subjectName, String subjectType) {
+    public Subject(String seq, String subjectName, SubjectType subjectType) {
         this.subjectId = seq;
         this.subjectName = subjectName;
         this.subjectType = subjectType;
@@ -22,7 +22,7 @@ public class Subject {
         return subjectName;
     }
 
-    public String getSubjectType() {
+    public SubjectType getSubjectType() {
         return subjectType;
     }
